chore: address requested changes
