+++ src/commands/tag.ts
@@ -3,7 +3,7 @@ import { ApplicationCommandOptionType, EmbedBuilder } from "discord.js";
 import { existsSync } from "fs";
 import { Paginator } from "../Paginator";
 import { publish } from "../plugins/publish";
-import type { TagData } from "./handlers/tagCreate";
+import type { TagData } from "../types";
 
 export default commandModule({
 	type: CommandType.Slash,
