+++ src/commands/tags.ts
@@ -13,7 +13,7 @@ import {
 import { existsSync, writeFileSync } from "fs";
 import { ownerOnly } from "../plugins/ownerOnly";
 import { publish } from "../plugins/publish";
-import type { TagData } from "./handlers/tagCreate";
+import type { TagData } from "../types";
 export default commandModule({
 	type: CommandType.Slash,
 	plugins: [publish(), ownerOnly()],
